+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- **httperr:** Fix wrapped HTTP error formatting
+
 
 ## [0.10.0] - 2018-08-21